requirements.txt
Flask==3.0.3
python-docx==1.1.2
pdfminer.six==20231228
beautifulsoup4==4.12.3
lxml==5.3.0
spacy==3.7.5
wordcloud==1.9.3